+++ packages/client/tsconfig.json
@@ -1,4 +1,4 @@
 {
 	"extends": "../../tsconfig.json",
-	"include": ["src"],
+	"include": ["src", "../test"],
 }
