+++ packages/common/src/index.ts
@@ -1,0 +1,3 @@
+export interface HELLO {
+  hello: string;
+}
