+++ packages/server/yarn-error.log
@@ -1,0 +1,129 @@
+Arguments: 
+  /home/hdn/.nvm/versions/node/v12.18.0/bin/node /usr/share/yarn/bin/yarn.js
+
+PATH: 
+  /tmp/yarn--1608921630518-0.6744613305407174:/home/hdn/dev/repos/au-lsp/node_modules/.bin:/home/hdn/.config/yarn/link/node_modules/.bin:/home/hdn/.yarn/bin:/home/hdn/.nvm/versions/node/v12.18.0/libexec/lib/node_modules/npm/bin/node-gyp-bin:/home/hdn/.nvm/versions/node/v12.18.0/lib/node_modules/npm/bin/node-gyp-bin:/home/hdn/.nvm/versions/node/v12.18.0/bin/node_modules/npm/bin/node-gyp-bin:/home/hdn/.local/bin:/home/hdn/.nvm/versions/node/v12.18.0/bin:/home/hdn/.local/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin
+
+Yarn version: 
+  1.21.1
+
+Node version: 
+  12.18.0
+
+Platform: 
+  linux x64
+
+Trace: 
+  Error: https://registry.yarnpkg.com/@aulsp%2fcommon: Not found
+      at Request.params.callback [as _callback] (/usr/share/yarn/lib/cli.js:66947:18)
+      at Request.self.callback (/usr/share/yarn/lib/cli.js:140665:22)
+      at Request.emit (events.js:315:20)
+      at Request.<anonymous> (/usr/share/yarn/lib/cli.js:141637:10)
+      at Request.emit (events.js:315:20)
+      at IncomingMessage.<anonymous> (/usr/share/yarn/lib/cli.js:141559:12)
+      at Object.onceWrapper (events.js:421:28)
+      at IncomingMessage.emit (events.js:327:22)
+      at endReadableNT (_stream_readable.js:1221:12)
+      at processTicksAndRejections (internal/process/task_queues.js:84:21)
+
+npm manifest: 
+  {
+  	"name": "lsp-sample-server",
+  	"description": "Example implementation of a language server in node.",
+  	"version": "1.0.0",
+  	"author": "Microsoft Corporation",
+  	"license": "MIT",
+  	"engines": {
+  		"node": "*"
+  	},
+  	"repository": {
+  		"type": "git",
+  		"url": "https://github.com/Microsoft/vscode-extension-samples"
+  	},
+  	"dependencies": {
+  		"parse5": "^6.0.1",
+  		"parse5-sax-parser": "^6.0.1",
+  		"vscode-languageserver": "^6.1.1",
+  		"vscode-languageserver-textdocument": "^1.0.1",
+  		"@aulsp/common": ""
+  	},
+  	"scripts": {},
+  	"devDependencies": {
+  		"@aurelia/kernel": "^0.6.0",
+  		"@types/parse5": "^5.0.3",
+  		"@types/parse5-sax-parser": "^5.0.1"
+  	}
+  }
+
+yarn manifest: 
+  No manifest
+
+Lockfile: 
+  # THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
+  # yarn lockfile v1
+  
+  
+  "@aurelia/kernel@^0.6.0":
+    version "0.6.0"
+    resolved "https://registry.yarnpkg.com/@aurelia/kernel/-/kernel-0.6.0.tgz#665e1dfd2e48e9ab51a3a64a92885b29a9f3cf1c"
+    integrity sha512-BbUecngdQt8D9xs2417llcdGUtVsiFgmpNeavFY5LkbrsJI3wWwphwenc8PCBqmSFxKJL2GnmPyav4e8xwYLFQ==
+  
+  "@types/node@*":
+    version "14.14.9"
+    resolved "https://registry.yarnpkg.com/@types/node/-/node-14.14.9.tgz#04afc9a25c6ff93da14deabd65dc44485b53c8d6"
+    integrity sha512-JsoLXFppG62tWTklIoO4knA+oDTYsmqWxHRvd4lpmfQRNhX6osheUOWETP2jMoV/2bEHuMra8Pp3Dmo/stBFcw==
+  
+  "@types/parse5-sax-parser@^5.0.1":
+    version "5.0.1"
+    resolved "https://registry.yarnpkg.com/@types/parse5-sax-parser/-/parse5-sax-parser-5.0.1.tgz#f1e26e82bb09e48cb0c16ff6d1e88aea1e538fd5"
+    integrity sha512-wBEwg10aACLggnb44CwzAA27M1Jrc/8TR16zA61/rKO5XZoi7JSfLjdpXbshsm7wOlM6hpfvwygh40rzM2RsQQ==
+    dependencies:
+      "@types/node" "*"
+      "@types/parse5" "*"
+  
+  "@types/parse5@*", "@types/parse5@^5.0.3":
+    version "5.0.3"
+    resolved "https://registry.yarnpkg.com/@types/parse5/-/parse5-5.0.3.tgz#e7b5aebbac150f8b5fdd4a46e7f0bd8e65e19109"
+    integrity sha512-kUNnecmtkunAoQ3CnjmMkzNU/gtxG8guhi+Fk2U/kOpIKjIMKnXGp4IJCgQJrXSgMsWYimYG4TGjz/UzbGEBTw==
+  
+  parse5-sax-parser@^6.0.1:
+    version "6.0.1"
+    resolved "https://registry.yarnpkg.com/parse5-sax-parser/-/parse5-sax-parser-6.0.1.tgz#98b4d366b5b266a7cd90b4b58906667af882daba"
+    integrity sha512-kXX+5S81lgESA0LsDuGjAlBybImAChYRMT+/uKCEXFBFOeEhS52qUCydGhU3qLRD8D9DVjaUo821WK7DM4iCeg==
+    dependencies:
+      parse5 "^6.0.1"
+  
+  parse5@^6.0.1:
+    version "6.0.1"
+    resolved "https://registry.yarnpkg.com/parse5/-/parse5-6.0.1.tgz#e1a1c085c569b3dc08321184f19a39cc27f7c30b"
+    integrity sha512-Ofn/CTFzRGTTxwpNEs9PP93gXShHcTq255nzRYSKe8AkVpZY7e1fpmTfOyoIvjP5HG7Z2ZM7VS9PPhQGW2pOpw==
+  
+  vscode-jsonrpc@^5.0.1:
+    version "5.0.1"
+    resolved "https://registry.yarnpkg.com/vscode-jsonrpc/-/vscode-jsonrpc-5.0.1.tgz#9bab9c330d89f43fc8c1e8702b5c36e058a01794"
+    integrity sha512-JvONPptw3GAQGXlVV2utDcHx0BiY34FupW/kI6mZ5x06ER5DdPG/tXWMVHjTNULF5uKPOUUD0SaXg5QaubJL0A==
+  
+  vscode-languageserver-protocol@^3.15.3:
+    version "3.15.3"
+    resolved "https://registry.yarnpkg.com/vscode-languageserver-protocol/-/vscode-languageserver-protocol-3.15.3.tgz#3fa9a0702d742cf7883cb6182a6212fcd0a1d8bb"
+    integrity sha512-zrMuwHOAQRhjDSnflWdJG+O2ztMWss8GqUUB8dXLR/FPenwkiBNkMIJJYfSN6sgskvsF0rHAoBowNQfbyZnnvw==
+    dependencies:
+      vscode-jsonrpc "^5.0.1"
+      vscode-languageserver-types "3.15.1"
+  
+  vscode-languageserver-textdocument@^1.0.1:
+    version "1.0.1"
+    resolved "https://registry.yarnpkg.com/vscode-languageserver-textdocument/-/vscode-languageserver-textdocument-1.0.1.tgz#178168e87efad6171b372add1dea34f53e5d330f"
+    integrity sha512-UIcJDjX7IFkck7cSkNNyzIz5FyvpQfY7sdzVy+wkKN/BLaD4DQ0ppXQrKePomCxTS7RrolK1I0pey0bG9eh8dA==
+  
+  vscode-languageserver-types@3.15.1:
+    version "3.15.1"
+    resolved "https://registry.yarnpkg.com/vscode-languageserver-types/-/vscode-languageserver-types-3.15.1.tgz#17be71d78d2f6236d414f0001ce1ef4d23e6b6de"
+    integrity sha512-+a9MPUQrNGRrGU630OGbYVQ+11iOIovjCkqxajPa9w57Sd5ruK8WQNsslzpa0x/QJqC8kRc2DUxWjIFwoNm4ZQ==
+  
+  vscode-languageserver@^6.1.1:
+    version "6.1.1"
+    resolved "https://registry.yarnpkg.com/vscode-languageserver/-/vscode-languageserver-6.1.1.tgz#d76afc68172c27d4327ee74332b468fbc740d762"
+    integrity sha512-DueEpkUAkD5XTR4MLYNr6bQIp/UFR0/IPApgXU3YfCBCB08u2sm9hRCs6DxYZELkk++STPjpcjksR2H8qI3cDQ==
+    dependencies:
+      vscode-languageserver-protocol "^3.15.3"
