+++ packages/tests/src/diagnostics.test.ts
