+++ packages/tests/src/helper.ts
@@ -13,9 +13,9 @@ import * as path from "path";
 import * as ts from "typescript";
 
 import { Container } from "aurelia-dependency-injection";
-import { AureliaProgram } from "../../../server/src/viewModel/AureliaProgram";
-import { createAureliaWatchProgram } from "../../../server/src/viewModel/createAureliaWatchProgram";
-import { setAureliaComponentMap } from "../../../server/src/viewModel/setAureliaComponentMap";
+import { AureliaProgram } from "../../server/src/viewModel/AureliaProgram";
+import { createAureliaWatchProgram } from "../../server/src/viewModel/createAureliaWatchProgram";
+import { setAureliaComponentMap } from "../../server/src/viewModel/setAureliaComponentMap";
 
 export let doc: vscode.TextDocument;
 export let editor: vscode.TextEditor;
