+++ packages/tests/src/testFixture/diagnostics.txt
